+++ src/pymeshlab/docs/pymesh_doc.h
@@ -61,6 +61,15 @@ const char* PYMESH_FN = "Returns the number of faces of the mesh.";
 // Mesh::edge_number
 const char* PYMESH_EN = "Returns the number of edges of the mesh.";
 
+// Mesh::id
+const char* PYMESH_ID =
+	"Returns the id of the mesh into a MeshSet. If the Mesh is not part of a MeshSet, the returned "
+	"value is -1.";
+
+// Mesh::label
+const char* PYMESH_LABEL =
+	"Returns a the label associated to the Mesh.";
+
 // Mesh::is_compact
 const char* PYMESH_IS_COMPACT = "Returns true if all its arrays are compact.";
 
@@ -72,6 +81,9 @@ const char* PYMESH_SEL_VN = "Returns the number of selected vertices of the mesh
 
 // Mesh::selected_face_number
 const char* PYMESH_SEL_FN = "Returns the number of selected faces of the mesh.";
+
+// Mesh::set_label
+const char* PYMESH_SET_LABEL = "Sets the label associated to the mesh.";
 
 // Mesh::update_bounding_box
 const char* PYMESH_UPDATE_BB = "Updates the bounding box of the mesh.";